Refactor file naming & other cmds
